move code to allow importing slice mixins without circular imports
