+++ jdaviz/configs/cubeviz/plugins/__init__.py
@@ -1,4 +1,5 @@
 from .tools import *  # noqa
+from .mixins import *  # noqa
 from .viewers import *  # noqa
 from .parsers import *  # noqa
 from .moment_maps.moment_maps import *  # noqa
