+++ jdaviz/configs/default/plugins/viewers.py
@@ -6,13 +6,12 @@ from glue_jupyter.bqplot.profile import BqplotProfileView
 from glue_jupyter.bqplot.image import BqplotImageView
 from glue_jupyter.table import TableViewer
 
-from jdaviz.configs.imviz.helper import layer_is_image_data
 from jdaviz.components.toolbar_nested import NestedJupyterToolbar
 from jdaviz.core.astrowidgets_api import AstrowidgetsImageViewerMixin
 from jdaviz.core.registries import viewer_registry
 from jdaviz.core.template_mixin import WithCache
 from jdaviz.core.user_api import ViewerUserApi
-from jdaviz.utils import ColorCycler, get_subset_type, _wcs_only_label, layer_is_not_dq
+from jdaviz.utils import ColorCycler, get_subset_type, _wcs_only_label, layer_is_image_data, layer_is_not_dq
 
 __all__ = ['JdavizViewerMixin']
