+++ jdaviz/utils.py
@@ -23,7 +23,9 @@ from ipyvue import watch
 
 __all__ = ['SnackbarQueue', 'enable_hot_reloading', 'bqplot_clear_figure',
            'standardize_metadata', 'ColorCycler', 'alpha_index', 'get_subset_type',
-           'download_uri_to_path', 'flux_conversion', 'spectral_axis_conversion']
+           'download_uri_to_path', 'flux_conversion', 'spectral_axis_conversion',
+           'layer_is_2d', 'layer_is_2d_or_3d', 'layer_is_image_data', 'layer_is_wcs_only',
+           'get_wcs_only_layer_labels', 'get_top_layer_index', 'get_reference_image_data']
 
 NUMPY_LT_2_0 = not minversion("numpy", "2.0.dev")
 
@@ -614,3 +616,67 @@ def download_uri_to_path(possible_uri, cache=None, local_path=os.curdir, timeout
 
     # assume this isn't a URI after all:
     return possible_uri
+
+def layer_is_2d(layer):
+    # returns True for subclasses of BaseData with ndim=2, both for
+    # layers that are WCS-only as well as images containing data:
+    return isinstance(layer, BaseData) and layer.ndim == 2
+
+
+def layer_is_2d_or_3d(layer):
+    return isinstance(layer, BaseData) and layer.ndim in (2, 3)
+
+
+def layer_is_image_data(layer):
+    return layer_is_2d_or_3d(layer) and not layer.meta.get(_wcs_only_label, False)
+
+
+def layer_is_wcs_only(layer):
+    return layer_is_2d(layer) and layer.meta.get(_wcs_only_label, False)
+
+
+def get_wcs_only_layer_labels(app):
+    return [data.label for data in app.data_collection
+            if layer_is_wcs_only(data)]
+
+
+def get_top_layer_index(viewer):
+    """Get index of the top visible image layer in a viewer.
+    This is because when blinked, first layer might not be top visible layer.
+
+    """
+    # exclude children of layer associations
+    associations = viewer.jdaviz_app._data_associations
+
+    visible_image_layers = [
+        i for i, lyr in enumerate(viewer.layers)
+        if (
+            lyr.visible and
+            layer_is_image_data(lyr.layer) and
+            # check that this layer is a root, without parents:
+            associations[lyr.layer.label]['parent'] is None
+        )
+    ]
+
+    if len(visible_image_layers):
+        return visible_image_layers[-1]
+    return None
+
+
+def get_reference_image_data(app, viewer_id=None):
+    """
+    Return the current reference data in the given image viewer and its index.
+    By default, the first viewer is used.
+    """
+    if viewer_id is None:
+        refdata = app._jdaviz_helper.default_viewer._obj.state.reference_data
+    else:
+        viewer = app.get_viewer_by_id(viewer_id)
+        refdata = viewer.state.reference_data
+
+    if refdata is not None:
+        iref = app.data_collection.index(refdata)
+        return refdata, iref
+
+    return None, -1
+
